+++ ccc-20200701/src/Models/ListRealtimeSkillGroupStatesRequest.php
@@ -18,6 +18,11 @@ class ListRealtimeSkillGroupStatesRequest extends Model
     public $instanceId;
 
     /**
+     * @var string
+     */
+    public $mediaType;
+
+    /**
      * @description This parameter is required.
      *
      * @example 1
@@ -43,6 +48,7 @@ class ListRealtimeSkillGroupStatesRequest extends Model
     public $skillGroupIdList;
     protected $_name = [
         'instanceId'       => 'InstanceId',
+        'mediaType'        => 'MediaType',
         'pageNumber'       => 'PageNumber',
         'pageSize'         => 'PageSize',
         'skillGroupIdList' => 'SkillGroupIdList',
@@ -57,6 +63,9 @@ class ListRealtimeSkillGroupStatesRequest extends Model
         $res = [];
         if (null !== $this->instanceId) {
             $res['InstanceId'] = $this->instanceId;
+        }
+        if (null !== $this->mediaType) {
+            $res['MediaType'] = $this->mediaType;
         }
         if (null !== $this->pageNumber) {
             $res['PageNumber'] = $this->pageNumber;
@@ -81,6 +90,9 @@ class ListRealtimeSkillGroupStatesRequest extends Model
         $model = new self();
         if (isset($map['InstanceId'])) {
             $model->instanceId = $map['InstanceId'];
+        }
+        if (isset($map['MediaType'])) {
+            $model->mediaType = $map['MediaType'];
         }
         if (isset($map['PageNumber'])) {
             $model->pageNumber = $map['PageNumber'];